[#19] Some small code optimizing
