+++ app/src/main/java/co/nimblehq/compose/crypto/ui/screens/home/HomeViewModel.kt
@@ -11,7 +11,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.*
 import javax.inject.Inject
 
-private const val MY_COINS_CURRENCY = "usd"
+const val FIAT_CURRENCY = "usd"
 private const val MY_COINS_ORDER = "market_cap_desc"
 private const val MY_COINS_PRICE_CHANGE_IN_HOUR = "24h"
 private const val MY_COINS_ITEM_PER_PAGE = 10
@@ -72,7 +72,7 @@ class HomeViewModel @Inject constructor(
         _showMyCoinsLoading.value = true
         getMyCoinsUseCase.execute(
             GetMyCoinsUseCase.Input(
-                currency = MY_COINS_CURRENCY,
+                currency = FIAT_CURRENCY,
                 order = MY_COINS_ORDER,
                 priceChangeInHour = MY_COINS_PRICE_CHANGE_IN_HOUR,
                 itemPerPage = MY_COINS_ITEM_PER_PAGE,
@@ -92,7 +92,7 @@ class HomeViewModel @Inject constructor(
         _showTrendingCoinsLoading.value = true
         getTrendingCoinsUseCase.execute(
             GetTrendingCoinsUseCase.Input(
-                currency = MY_COINS_CURRENCY,
+                currency = FIAT_CURRENCY,
                 order = MY_COINS_ORDER,
                 priceChangeInHour = MY_COINS_PRICE_CHANGE_IN_HOUR,
                 itemPerPage = MY_COINS_ITEM_PER_PAGE,
