+++ app/src/main/java/co/nimblehq/compose/crypto/ui/uimodel/CoinDetailUiModel.kt
@@ -2,9 +2,8 @@ package co.nimblehq.compose.crypto.ui.uimodel
 
 import co.nimblehq.compose.crypto.data.extension.orZero
 import co.nimblehq.compose.crypto.domain.model.CoinDetail
+import co.nimblehq.compose.crypto.ui.screens.home.FIAT_CURRENCY
 import java.math.BigDecimal
-
-private const val COIN_CURRENCY_KEY = "usd"
 
 data class CoinDetailUiModel(
     val coinName: String,
@@ -22,13 +21,13 @@ data class CoinDetailUiModel(
 fun CoinDetail.toUiModel() = CoinDetailUiModel(
     coinName = coinName,
     image = image?.small.orEmpty(),
-    currentPrice = marketData?.currentPrice?.getValue(COIN_CURRENCY_KEY).orZero(),
+    currentPrice = marketData?.currentPrice?.getValue(FIAT_CURRENCY).orZero(),
     priceChangePercentage24hInCurrency = marketData?.priceChangePercentage24hInCurrency
-        ?.getValue(COIN_CURRENCY_KEY).orZero(),
-    marketCap = marketData?.marketCap?.getValue(COIN_CURRENCY_KEY).orZero(),
+        ?.getValue(FIAT_CURRENCY).orZero(),
+    marketCap = marketData?.marketCap?.getValue(FIAT_CURRENCY).orZero(),
     marketCapChangePercentage24h = marketData?.marketCapChangePercentage24h.orZero(),
-    ath = marketData?.ath?.getValue(COIN_CURRENCY_KEY).orZero(),
-    athChangePercentage = marketData?.athChangePercentage?.getValue(COIN_CURRENCY_KEY).orZero(),
-    atl = marketData?.atl?.getValue(COIN_CURRENCY_KEY).orZero(),
-    atlChangePercentage = marketData?.atlChangePercentage?.getValue(COIN_CURRENCY_KEY).orZero()
+    ath = marketData?.ath?.getValue(FIAT_CURRENCY).orZero(),
+    athChangePercentage = marketData?.athChangePercentage?.getValue(FIAT_CURRENCY).orZero(),
+    atl = marketData?.atl?.getValue(FIAT_CURRENCY).orZero(),
+    atlChangePercentage = marketData?.atlChangePercentage?.getValue(FIAT_CURRENCY).orZero()
 )
